refactor: rename Navigation_v2 to Navigation
