+++ layout/DefaultLayout.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navigation from '@/shared/components/Navigation_v2';
+import Navigation from '@/shared/components/Navigation';
 import Footer from '@/shared/components/Footer_v2';
 import { NavigationProvider } from '@/contexts/Navigation';
 import { PromotionProvider } from '@/contexts/Promotion';
