+++ pages/partner/index.jsx
@@ -5,7 +5,7 @@ import { NavigationProvider } from '@/contexts/Navigation';
 import { PromotionProvider } from '@/contexts/Promotion';
 import SEOConfig from '../../shared/components/SEO';
 import Partner from '../../components/Partner';
-import Navigation from '../../shared/components/Navigation_v2';
+import Navigation from '../../shared/components/Navigation';
 import Footer from '../../shared/components/Footer_v2';
 
 const HomePageWrapper = styled.div`
