+++ shared/components/Navigation/MainNav/Logo.jsx
