+++ shared/components/Navigation/MainNav/SubList/UserAvatar/index.jsx
