+++ shared/components/Navigation/PromotionBar.jsx
